+++ demo/app/app.ts
@@ -1,5 +1,5 @@
-﻿import * as app from 'application';
-import * as trace from 'trace';
+﻿import * as app from 'tns-core-modules/application';
+import * as trace from 'tns-core-modules/trace';
 
 app.on(app.launchEvent, () => {
     trace.addCategories(`${trace.categories.Debug},${trace.categories.Error}`)